% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullets}
\alias{bullets}
\title{Information of two example bullets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.
}
\source{
\url{https://sctyner.github.io/OpenForSciR/bullets.html}
}
\usage{
bullets
}
\description{
A dataset containing pre-processed information of two bullets. They are used
as examples in Chapter 3.5 of Open Forensic Science in R.
}
\keyword{datasets}
