% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccf4.R
\name{cmps_na_trim}
\alias{cmps_na_trim}
\title{Remove the leading and trailing missing values in a numeric vector}
\usage{
cmps_na_trim(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector; only the leading and trailing missing values are removed
}
\description{
Remove the leading and trailing missing values in a numeric vector
}
\examples{
x <- c(NA, 1, 2, 3, 4, NA)
cmps_na_trim(x)
}
