% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_ss_ratio}
\alias{compute_ss_ratio}
\title{#' Compute the Sum of Squares Ratio}
\usage{
compute_ss_ratio(score, label, MS = FALSE)
}
\arguments{
\item{score}{a numeric vector, scores}

\item{label}{a character vector, the label of each score}

\item{MS}{boolean, whether to compute the mean squares instead of the sum of squares. Default is FALSE}
}
\value{
the sum of squares ratio
}
\description{
#' Compute the Sum of Squares Ratio
}
\examples{
score <- c(rnorm(100), rnorm(100, mean = 5))
label <- c(rep("a", 100), rep("b", 100))
compute_ss_ratio(score, label)
}
