% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccf3.R
\name{get_ccf3}
\alias{get_ccf3}
\title{Function to calculate the cross-correlation between two sequences}
\usage{
get_ccf3(x, y, min.overlap = round(0.1 * max(length(x), length(y))))
}
\arguments{
\item{x}{numeric sequence of values}

\item{y}{numeric sequence of values}

\item{min.overlap}{integer, minimal number of values in the overlap between sequences x and y to calculate a correlation value. Set to 10 percent of the maximum length of either sequence (HH: this might be problematic for CMPS)}
}
\value{
list consisting of the lag where the maximum correlation is achieved, and the maximum correlation value.
}
\description{
some more description, x is the longer (reference) sequence
}
