% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{get_ccp}
\alias{get_ccp}
\title{Identify at most one consistent correlation peak (ccp)}
\usage{
get_ccp(ccr.list, Tx = 25)
}
\arguments{
\item{ccr.list}{list, obtained by \code{get_ccr_peaks}}

\item{Tx}{integer, the tolerance zone is \verb{+/- Tx}}
}
\value{
integer, the position of the ccp if it is identified; \code{NULL} otherwise.
}
\description{
If multi segment lengths strategy is being used, at most one consistent correlation
peak (ccp) will be found for the corresponding basis segment. If the ccp cannot be identified,
return \code{NULL}
}
