% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{get_ccr_peaks}
\alias{get_ccr_peaks}
\title{Identify peaks of a cross correlation curve}
\usage{
get_ccr_peaks(comp, segments, seg_scale, nseg = 1, npeaks = 5)
}
\arguments{
\item{comp}{a nueric vector, vector of the bullet comparison profile}

\item{segments}{list with basis segments and their corresponding indices in the original profile, obtianed by \code{get_segs()}}

\item{seg_scale}{integer. the scale of the segment.
\itemize{
\item \code{seg_scale} = 1: the length remains the same;
\item \code{seg_scale} = 2: the length will be doubled.
\item \code{seg_scale} = 3: the length will be doubled twice.
}}

\item{nseg}{integer. \code{nseg} = 3: the third segment in \code{segments}}

\item{npeaks}{integer. the number of peaks to be identified.}
}
\value{
a list consisting of:
\itemize{
\item \code{ccr}: the cross correlation curve
\item \code{adj.pos}: indices of the curve
\item \code{peaks.pos}: position of the identified peaks
\item \code{peaks.heights}: the cross correlation value (height of the curve) of the peaks
}
}
\description{
Given a comparison profile and a segment, \code{get_ccr_peaks} computes the
cross correlation curve and finds peaks of the curve.
}
