% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{local_max_cmps}
\alias{local_max_cmps}
\title{find local maximums}
\usage{
local_max_cmps(x, find_max = 0)
}
\arguments{
\item{x}{numeric vector, the input sequence}

\item{find_max}{a numeric scalor, the function finds maximums if \code{find_max = 0}
finds minimums if overwise.}
}
\description{
find local maximums
}
