% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{metric_plot_helper}
\alias{metric_plot_helper}
\title{Helper Function for Plotting the Distribution of a Metric}
\usage{
metric_plot_helper(cmps.metric, metric, scaled = FALSE, SSratio = TRUE, ...)
}
\arguments{
\item{cmps.metric}{a data frame containing values of the metric and the labels}

\item{metric}{string. Which metric to be plotted}

\item{scaled}{logical value. If \code{scaled = TRUE}, the values should be within the interval of \verb{[0, 1]}}

\item{SSratio}{logical value. Whether to show the sum of squares ratio value}

\item{...}{other arguments for plotting: \code{breaks}, \code{binwidth}, and \code{subtitle}}
}
\value{
a ggplot object
}
\description{
Helper Function for Plotting the Distribution of a Metric
}
