% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullets}
\alias{bullets}
\title{Information of two example bullets}
\format{
A data frame/tbl/tbl_df with 12 rows and 3 variables:
\describe{
\item{source}{source of the bullet data}
\item{sigs}{bullet signatures, detailed information about how to get the signatures can be found at https://sctyner.github.io/OpenForSciR/bullets.html}
\item{bulletland}{label of the signatures}
}
}
\source{
\url{https://sctyner.github.io/OpenForSciR/bullets.html}
}
\usage{
bullets
}
\description{
A dataset containing pre-processed information of two bullets. They are used
as examples in Chapter 3.5 of Open Forensic Science in R.
}
\keyword{datasets}
