% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmps.R
\name{get_CMPS}
\alias{get_CMPS}
\title{Compute the CMPS score}
\usage{
get_CMPS(input.ccp, Tx = 25, order = T)
}
\arguments{
\item{input.ccp}{a list of positions for (consistent) correlation peaks}

\item{Tx}{integer, the tolerance zone is \verb{+/- Tx}}

\item{order}{boolean, whether or not to order peak positions based on their CMPS score}
}
\value{
a list of three components:
\itemize{
\item \code{CMPS.score}: computed CMPS score
\item \code{rec.position}: the recommended position that results in the CMPS score
\item \code{pos.df}: a dataframe that includes all positions and their corresponding CMPS score
}
}
\description{
Compute the CMPS score from a list of positions of (consistent) correlation peaks.
}
